# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Web Server Basic Template API",
    "version": "1.0.0",
    "description": "API di esempio per un server Express con validazione Zod. Risorse: users, teachers, classes, rubrica. I dati sono in memoria: lo stato si azzera a ogni riavvio del server.",
    "contact": {
      "name": "Web Server Basic Template"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Server di sviluppo locale"
    }
  ],
...
  "paths": {
...
    "/api/rubrica": {
      "get": {
        "tags": ["Rubrica"],
        "summary": "List all contacts",
        "description": "Restituisce la lista completa dei contatti della rubrica.",
        "operationId": "getRubrica",
        "responses": {
          "200": {
            "description": "Lista dei contatti",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rubrica"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Rubrica"],
        "summary": "Create a contact",
        "description": "Aggiunge un nuovo contatto alla rubrica. Tutti i campi sono obbligatori.",
        "operationId": "createRubrica",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRubricaInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contatto creato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rubrica"
                }
              }
            }
          },
          "400": {
            "description": "Validazione fallita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rubrica/search": {
      "get": {
        "tags": ["Rubrica"],
        "summary": "Search a contact by email",
        "description": "Cerca un contatto nella rubrica in base all'email. Se nessun contatto corrisponde restituisce un oggetto vuoto.",
        "operationId": "searchRubrica",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "description": "Email del contatto da cercare",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contatto trovato (o oggetto vuoto se assente)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Rubrica"
                    },
                    {
                      "type": "object",
                      "example": {}
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/rubrica/{id}": {
      "get": {
        "tags": ["Rubrica"],
        "summary": "Get a contact by id",
        "description": "Restituisce un singolo contatto della rubrica identificato dal suo id numerico.",
        "operationId": "getRubricaById",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Contatto trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rubrica"
                }
              }
            }
          },
          "400": {
            "description": "Validazione fallita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contatto non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Identificativo numerico della risorsa",
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      }
    },
    "schemas": {
...
      "Rubrica": {
        "type": "object",
        "description": "Contatto della rubrica restituito dalle API",
        "required": ["id", "name", "surname", "email", "telephone"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificativo univoco",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Mario"
          },
          "surname": {
            "type": "string",
            "example": "Rossi"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "mario.rossi@example.com"
          },
          "telephone": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10,
            "example": "1234567890"
          }
        }
      },
      "CreateRubricaInput": {
        "type": "object",
        "description": "Payload per l'aggiunta di un contatto in rubrica",
        "required": ["name", "surname", "email", "telephone"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "Mario"
          },
          "surname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "Rossi"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "mario.rossi@example.com"
          },
          "telephone": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10,
            "example": "1234567890"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Risposta di errore standard",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "description": "Dettagli degli errori di validazione (presenti solo con status 400)",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Invalid email address"
                }
              }
            }
          }
        }
      }
    }
  }
}